do not include tag prefix in version
If tag is prefixed (e.g. `v1.0.0`) the prefix should not be considered
as part of version string. Therefore a check is added for tagged commits
if the tag includes version in package.json. If so everything else from tag
is treated as a prefix (or suffix). If not tag name is returned as before.

Some examples:
tag             version in package.json     returned
v1.0.0          1.0.0                       1.0.0
release/1.0.0   1.0.0                       1.0.0
v1.0.0          - none -                    v0.1.0
bar             foo                         bar

Fixes #11